+++ illumio-core/resource_illumio_service.go
@@ -386,7 +386,7 @@ func resourceIllumioServiceRead(ctx context.Context, d *schema.ResourceData, m i
 
 	_, data, err := illumioClient.Get(d.Id(), nil)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleReadError(err, d, "illumio-core_service")
 	}
 
 	for _, key := range []string{