+++ illumio-core/resource_illumio_service_binding.go
@@ -223,7 +223,7 @@ func resourceIllumioServiceBindingRead(ctx context.Context, d *schema.ResourceDa
 
 	_, data, err := illumioClient.Get(href, nil)
 	if err != nil {
-		return diag.FromErr(err)
+		return handleReadError(err, d, "illumio-core_service_binding")
 	}
 
 	d.SetId(data.S("href").Data().(string))